Proposal for conditionals 3: parfor outside
This is the complete opposite to proposal 1 (#15). The parfor loop is completely outside of the if.
Compared to the current GTScript, the `_INLINE if` is changed to a `with condition()`, because ti would the following would be weird:
```
if True:
    b = a
    c = b
```
and
```
if _INLINE(True):
    b = a
    c = b
```
means different execution which is very confusing.

Note that this proposal has a clear downside for large if blocks as everything is one big parfor loop, e.g. in the following case
```python
with computation(...) and stuff:
    if runtime_flag:
        # a lot of code here
        tmp = a[-1,0,0]
        # a lot of code here
        a = tmp[-1,0,0]
```
This is Eddies famous example, which currently fails but (without the if) should actually be legal. In this proposal, the code would have to be split because of the race condition to something like
```python
with computation(...) and stuff:
    if runtime_flag:
        # a lot of code here
        tmp = a[-1,0,0]
        # a lot of code here
    if runtime_flag:
        a = tmp[-1,0,0]
```
Now add an else or some nested `if`s and you get problems.

TODO: more examples